fixing promoting video bug for android
